+++ assets/plugins/wwsu-remote/js/wwsu-remote.js
@@ -4,11 +4,13 @@ class WWSUremote extends WWSUevents {
 	 *
 	 * @param {sails.io} socket Socket connection to WWSU
 	 * @param {WWSUreq} hostReq Request with host authorization
+	 * @param {WWSUrecipients} recipients Initialized recipients module (uses recipients.recipient for Skyway credential authorization and peerId)
 	 */
-	constructor(socket, hostReq) {
+	constructor(socket, hostReq, recipients) {
 		super();
 		this.endpoints = {
 			request: "/call/request",
+			credentialComputer: "/call/credential-computer",
 		};
 		this.requests = {
 			host: hostReq,
@@ -16,6 +18,8 @@ class WWSUremote extends WWSUevents {
 		this.data = {
 			request: {},
 		};
+
+		this.recipients = recipients;
 	}
 
 	/**
@@ -55,6 +59,55 @@ class WWSUremote extends WWSUevents {
 				title: "Error requesting audio call",
 				body:
 					"There was an error informing the WWSU API we want to start an audio call for a remote broadcast. Please contact the engineer.",
+				autoHide: true,
+				delay: 15000,
+				icon: "fas fa-skull-crossbones fa-lg",
+			});
+			if (typeof cb === "function") {
+				cb(false);
+			}
+			console.error(e);
+		}
+	}
+
+	/**
+	 * Authorize a host for connecting to Skyway.js.
+	 * TODO: Once the new DJ Controls is ready, credentials should be forced on the Skyway.js dashboard.
+	 * 
+	 * @param {object} data Data to be passed to the API
+	 * @param {?function} cb Callback which returns credential data on success
+	 */
+	credentialComputer(data, cb) {
+		try {
+			this.requests.host.request(
+				{ method: "post", url: this.endpoints.credentialComputer, data },
+				(response) => {
+					if (response.authToken) {
+						if (typeof cb === "function") {
+							cb(response);
+						}
+					} else {
+						$(document).Toasts("create", {
+							class: "bg-danger",
+							title: "Error generating Skyway.js credential",
+							body:
+								"There was an error generating a credential token to authorize Skyway.js for an audio call. Please contact the engineer.",
+							autoHide: true,
+							delay: 15000,
+							icon: "fas fa-skull-crossbones fa-lg",
+						});
+						if (typeof cb === "function") {
+							cb(false);
+						}
+					}
+				}
+			);
+		} catch (e) {
+			$(document).Toasts("create", {
+				class: "bg-danger",
+				title: "Error generating Skyway.js credential",
+				body:
+					"There was an error generating a credential token to authorize Skyway.js for an audio call. Please contact the engineer.",
 				autoHide: true,
 				delay: 15000,
 				icon: "fas fa-skull-crossbones fa-lg",